requirements.txt
blinker==1.8.2
click==8.1.8
flask==3.0.3
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
pillow==10.4.0
PyBluez==0.23
pypng==0.20220715.0
pystray==0.19.5
python-xlib==0.33
qrcode==7.4.2
six==1.17.0
typing-extensions==4.13.2
werkzeug==3.0.6
zipp==3.20.2
